+++ packages/cosmic-swingset/economy-template.json
@@ -1,5 +1,6 @@
 [
   "@agoric/vats/scripts/init-core.js",
+  "@agoric/vats/scripts/init-network.js",
   {
     "module": "@agoric/inter-protocol/scripts/init-core.js",
     "entrypoint": "defaultProposalBuilder",
