+++ packages/vats/decentral-core-config.json
@@ -29,14 +29,8 @@
     "bridge": {
       "sourceSpec": "@agoric/vats/src/vat-bridge.js"
     },
-    "ibc": {
-      "sourceSpec": "@agoric/vats/src/vat-ibc.js"
-    },
     "mints": {
       "sourceSpec": "@agoric/vats/src/vat-mints.js"
-    },
-    "network": {
-      "sourceSpec": "@agoric/vats/src/vat-network.js"
     },
     "priceAuthority": {
       "sourceSpec": "@agoric/vats/src/vat-priceAuthority.js"
