+++ packages/vats/decentral-demo-config.json
@@ -5,6 +5,7 @@
   "snapshotInterval": 1000,
   "coreProposals": [
     "@agoric/vats/scripts/init-core.js",
+    "@agoric/vats/scripts/init-network.js",
     {
       "module": "@agoric/inter-protocol/scripts/init-core.js",
       "entrypoint": "defaultProposalBuilder",
@@ -125,12 +126,6 @@
     },
     "board": {
       "sourceSpec": "@agoric/vats/src/vat-board.js"
-    },
-    "ibc": {
-      "sourceSpec": "@agoric/vats/src/vat-ibc.js"
-    },
-    "network": {
-      "sourceSpec": "@agoric/vats/src/vat-network.js"
     },
     "priceAuthority": {
       "sourceSpec": "@agoric/vats/src/vat-priceAuthority.js"
