+++ packages/vats/decentral-devnet-config.json
@@ -4,6 +4,7 @@
   "defaultReapInterval": 1000,
   "coreProposals": [
     "@agoric/vats/scripts/init-core.js",
+    "@agoric/vats/scripts/init-network.js",
     {
       "module": "@agoric/inter-protocol/scripts/init-core.js",
       "entrypoint": "defaultProposalBuilder",
@@ -175,17 +176,11 @@
     "centralSupply": {
       "sourceSpec": "@agoric/vats/src/centralSupply.js"
     },
-    "ibc": {
-      "sourceSpec": "@agoric/vats/src/vat-ibc.js"
-    },
     "mintHolder": {
       "sourceSpec": "@agoric/vats/src/mintHolder.js"
     },
     "mints": {
       "sourceSpec": "@agoric/vats/src/vat-mints.js"
-    },
-    "network": {
-      "sourceSpec": "@agoric/vats/src/vat-network.js"
     },
     "priceAuthority": {
       "sourceSpec": "@agoric/vats/src/vat-priceAuthority.js"
