+++ packages/vats/demo-proposals.json
@@ -1,4 +1,5 @@
 [
   "@agoric/inter-protocol/scripts/init-core.js",
-  "@agoric/pegasus/scripts/init-core.js"
+  "@agoric/pegasus/scripts/init-core.js",
+  "@agoric/vats/scripts/init-network.js"
 ]
