+++ packages/vats/src/core/client-behaviors.js
@@ -3,7 +3,7 @@ import { E, Far } from '@endo/far';
 import { makePluginManager } from '@agoric/swingset-vat/src/vats/plugin-manager.js';
 import { observeNotifier } from '@agoric/notifier';
 import { deeplyFulfilledObject } from '@agoric/internal';
-import { registerNetworkProtocols } from './chain-behaviors.js';
+import { registerNetworkProtocols } from '../proposals/network-proposal.js';
 import { makeVatsFromBundles } from './basic-behaviors.js';
 
 const { Fail } = assert;
