+++ packages/vats/test/snapshots/test-boot-config.js.md
@@ -44,7 +44,6 @@ Generated by [AVA](https://avajs.dev).
     "space_zoe" [label="zoe", shape=house, style=filled, fillcolor=khaki];␊
     "space_zoe" [label="zoe", shape=house, style=filled, fillcolor=khaki];␊
     "installation_contractGovernor" [label="contractGovernor", shape=cylinder];␊
-    "space_zoe" [label="zoe", shape=house, style=filled, fillcolor=khaki];␊
     }␊
     subgraph cluster_board {␊
     label = "board"␊
@@ -56,7 +55,6 @@ Generated by [AVA](https://avajs.dev).
     "home" [label="home"];␊
     "home" [label="home"];␊
     "home" [label="home"];␊
-    "space_provisioning" [label="provisioning", shape=house, style=filled, fillcolor=khaki];␊
     "space_provisioning" [label="provisioning", shape=house, style=filled, fillcolor=khaki];␊
     "home" [label="home"];␊
     }␊
@@ -84,7 +82,6 @@ Generated by [AVA](https://avajs.dev).
     "space_bridgeManager" [label="bridgeManager", shape=house, style=filled, fillcolor=khaki];␊
     "space_provisionBridgeManager" [label="provisionBridgeManager", shape=house, style=filled, fillcolor=khaki];␊
     "space_provisionWalletBridgeManager" [label="provisionWalletBridgeManager", shape=house, style=filled, fillcolor=khaki];␊
-    "space_bridgeManager" [label="bridgeManager", shape=house, style=filled, fillcolor=khaki];␊
     }␊
     subgraph cluster_BLD {␊
     label = "BLD"␊
@@ -110,10 +107,6 @@ Generated by [AVA](https://avajs.dev).
     subgraph cluster_vattp {␊
     label = "vattp"␊
     "vats_vattp" [label="vattp", shape=doubleoctagon, style=filled, fillcolor=tomato];␊
-    }␊
-    subgraph cluster_network {␊
-    label = "network"␊
-    "space_networkVat" [label="networkVat", shape=house, style=filled, fillcolor=khaki];␊
     }␊
     "makeOracleBrands" [label="makeOracleBrands"];␊
     "oracleBrand_USD" [label="USD"];␊
@@ -216,11 +209,6 @@ Generated by [AVA](https://avajs.dev).
     "setupClientManager" [label="setupClientManager"];␊
     "space_client" [label="client", shape=house, style=filled, fillcolor=khaki];␊
     "space_clientCreator" [label="clientCreator", shape=house, style=filled, fillcolor=khaki];␊
-    "setupNetworkProtocols" [label="setupNetworkProtocols"];␊
-    "space_networkVat" [label="networkVat", shape=house, style=filled, fillcolor=khaki];␊
-    "space_bridgeManager" [label="bridgeManager", shape=house, style=filled, fillcolor=khaki];␊
-    "space_zoe" [label="zoe", shape=house, style=filled, fillcolor=khaki];␊
-    "space_provisioning" [label="provisioning", shape=house, style=filled, fillcolor=khaki];␊
     "connectChainFaucet" [label="connectChainFaucet"];␊
     "home_faucet" [label="faucet", shape=folder];␊
     "home" [label="home"];␊
@@ -306,10 +294,6 @@ Generated by [AVA](https://avajs.dev).
     "bridgeProvisioner" -> "space_provisionWalletBridgeManager" [dir=back]␊
     "setupClientManager" -> "space_client" []␊
     "setupClientManager" -> "space_clientCreator" []␊
-    "setupNetworkProtocols" -> "space_networkVat" []␊
-    "setupNetworkProtocols" -> "space_bridgeManager" [dir=back]␊
-    "setupNetworkProtocols" -> "space_zoe" [dir=back]␊
-    "setupNetworkProtocols" -> "space_provisioning" [dir=back]␊
     "connectChainFaucet" -> "home_faucet" []␊
     }␊
     `
